Updated gradle wrapper Updated gradle-intellij-plugin Added ManifestClasspath plugin for Windows builds and very long classpath Fixed inspection tests to run with IntelliJ 2019.1.EAP
