+++ gradle.properties
@@ -6,7 +6,8 @@
 #ideaVersion=2018.1.6
 #ideaVersion=2018.2
 #ideaVersion=2018.2.6
-ideaVersion=2018.3.2
+ideaVersion=2018.3.4
+#ideaVersion=191.5109.14
 #ideaVersion=LATEST-EAP-SNAPSHOT
 #ideaVersion = LATEST-TRUNK-SNAPSHOT
 #
@@ -21,5 +22,3 @@ descriptionFile=parts/pluginDescription.html
 changesFile=parts/pluginChanges.html
 #
 org.gradle.jvmargs=-XX:+CMSClassUnloadingEnabled -XX:+CMSPermGenSweepingEnabled -XX:+HeapDumpOnOutOfMemoryError -Xmx1024m -Dfile.encoding=utf-8
-#
-#localIdeaDependencyCachePath=D:/IdeaCache
