+++ testData/inspection/diverse/EqualsAndHashCodeCallSuper.java
@@ -1,11 +1,11 @@
-public class Test {
+public class EqualsAndHashCodeCallSuper {
 
   @lombok.EqualsAndHashCode
   private static class First {
     private String field1;
   }
 
-  @lombok.EqualsAndHashCode
+  <warning descr="Generating equals/hashCode implementation but without a call to superclass, even though this class does not extend java.lang.Object. If this is intentional, add '(callSuper=false)' to your type.">@lombok.EqualsAndHashCode</warning>
   private static class Second1 extends First {
     private String field2;
   }
@@ -22,13 +22,13 @@ public class Test {
 
 ///////////////////Data///////////////////////////
 
-  @lombok.Data
+  <warning descr="Generating equals/hashCode implementation but without a call to superclass, even though this class does not extend java.lang.Object. If this is intentional, add '(callSuper=false)' to your type.">@lombok.Data</warning>
   private static class Data1 extends First {
     private String field2;
   }
 
   @lombok.Data
-  @lombok.EqualsAndHashCode
+  <warning descr="Generating equals/hashCode implementation but without a call to superclass, even though this class does not extend java.lang.Object. If this is intentional, add '(callSuper=false)' to your type.">@lombok.EqualsAndHashCode</warning>
   private static class Data2 extends First {
     private String field2;
   }
@@ -47,13 +47,13 @@ public class Test {
 
 ///////////////////Value///////////////////////////
 
-  @lombok.Value
+  <warning descr="Generating equals/hashCode implementation but without a call to superclass, even though this class does not extend java.lang.Object. If this is intentional, add '(callSuper=false)' to your type.">@lombok.Value</warning>
   private static class Value1 extends First {
     private String field2;
   }
 
   @lombok.Value
-  @lombok.EqualsAndHashCode
+  <warning descr="Generating equals/hashCode implementation but without a call to superclass, even though this class does not extend java.lang.Object. If this is intentional, add '(callSuper=false)' to your type.">@lombok.EqualsAndHashCode</warning>
   private static class Value2 extends First {
     private String field2;
   }
