+++ testData/inspection/equalsAndHashCodeCallSuperConfigSkip/Test.java
