--- testData/inspection/equalsAndHashCodeCallSuperConfigSkip/expected.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<problems>
-
-</problems>
