+++ testData/inspection/equalsAndHashCodeCallSuperConfigSkip/lombok.config
