--- testData/inspection/valInspection/expected.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<problems>
-  <problem>
-    <file>Test.java</file>
-    <line>15</line>
-    <entry_point TYPE="method" FQNAME="valInspection void test()"/>
-    <problem_class severity="WARNING" attribute_key="WARNING_ATTRIBUTES">Lombok annotations inspection</problem_class>
-    <description>'val' is not allowed in old-style for loops</description>
-  </problem>
-
-  <problem>
-    <file>Test.java</file>
-    <line>20</line>
-    <entry_point TYPE="method" FQNAME="valInspection void test()"/>
-    <problem_class severity="WARNING" attribute_key="WARNING_ATTRIBUTES">Lombok annotations inspection</problem_class>
-    <description>'val' is not compatible with array initializer expressions. Use the full form (new int[] { ... } instead of just { ... })</description>
-  </problem>
-
-  <problem>
-    <file>Test.java</file>
-    <line>23</line>
-    <entry_point TYPE="method" FQNAME="valInspection void test()"/>
-    <problem_class severity="WARNING" attribute_key="WARNING_ATTRIBUTES">Lombok annotations inspection</problem_class>
-    <description>'val' is not allowed with lambda expressions.</description>
-  </problem>
-
-  <problem>
-    <file>Test.java</file>
-    <line>26</line>
-    <entry_point TYPE="method" FQNAME="valInspection void test()"/>
-    <problem_class severity="WARNING" attribute_key="WARNING_ATTRIBUTES">Lombok annotations inspection</problem_class>
-    <description>'val' on a local variable requires an initializer expression</description>
-  </problem>
-
-  <problem>
-    <file>Test.java</file>
-    <line>33</line>
-    <entry_point TYPE="method" FQNAME="valInspection void test(val x)"/>
-    <problem_class severity="WARNING" attribute_key="WARNING_ATTRIBUTES">Lombok annotations inspection</problem_class>
-    <description>'val' works only on local variables and on foreach loops</description>
-  </problem>
-</problems>
